+++ src/App.tsx
@@ -10,7 +10,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import React from 'react';
 import FileUploader from './components/FileUploader';
-import navData from './navData';
+import navData from './data/navData';
 
 const GridBG = React.lazy(() => import('./components/GridBG'));
