+++ src/Home.tsx
@@ -1,5 +1,5 @@
 import Article from "./components/Article";
-import { GridProps } from "./types";
+import { GridProps } from "./data/types";
 
 function App(gridParams: GridProps) {
   return <Article centreDotHandler={gridParams.centreDotHandler} dotSizeHandler={gridParams.dotSizeHandler} gridSizeHandler={gridParams.gridSizeHandler}/>;
