+++ src/data/jobTypes.ts
@@ -1,0 +1,21 @@
+interface JobParameter {
+  name: string,
+  dataType: string,
+  range?: [number, number]
+};
+
+export default [
+  {
+    name: "Greyscale"
+  },
+  {
+    name: "Kuwahara",
+    parameters: [
+      {
+        name: "Size",
+        dataType: "number",
+        range: [2, 32]
+      }
+    ]
+  }
+];
